+++ Sources/ProjectDescription/DependencyManager.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-public extension Dependency {
-    enum Manager: String, Codable, Equatable {
-        case carthage
-//        case spm
-//        case cocoapods
-    }
-}