+++ Sources/TuistSupport/Constants.swift
@@ -27,6 +27,13 @@ public struct Constants {
         public static let swiftDoc = "swift-doc"
     }
 
+    public struct DependenciesDirectory {
+        public static let name = "Dependencies"
+        public static let lockfilesDirectoryName = "Lockfiles"
+        public static let cartfileResolvedName = "Cartfile.resolved"
+        public static let carthageDirectoryName = "Carthage"
+    }
+
     public struct DerivedDirectory {
         public static let name = "Derived"
         public static let infoPlists = "InfoPlists"